+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #114 JUnit 4 support deprecation
 * #115 Table: loss of precision when inserting BigDecimal to Derby
+* #116 CsvDataSetBuilder can't handle BigDecimal
 
 ## 2.0.B1